cambios en servicios y jpa

+++ java/com/gestorDocumental/repository/ClienteRepositorio.java
@@ -9,7 +9,7 @@ import com.gestorDocumental.entity.Cliente;
 	
 public interface ClienteRepositorio extends JpaRepository<Cliente,Serializable>{
 	public abstract Cliente findByRazonSocial(String razonSocial);
-	public abstract Cliente findByNumeroCliente(Integer nroCliente);
+	public abstract Cliente findByNumeroCliente(long nroCliente);
 	
 	@Query("SELECT f.numeroCliente FROM Cliente f")
 	public abstract List<Object> obtenerNumerosCliente();
